requirements.txt
aiobalaboba==3.0.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pip==23.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
setuptools==67.6.0
urllib3==1.26.14
wheel==0.38.4
yarl==1.8.2
wcwidth==0.2.6
prettytable==3.6.0
openai==0.27.1
aiofiles==23.1.0
geocoder==1.38.1
future==0.18.3
click==8.1.3
decorator==5.1.1
six==1.16.0
ratelim==0.1.6